+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Upcoming
 
 - feat: add `memoization` option to `autorun`, default is `True`, compatible with old behavior, if set to `False`, calling the function explicitly will always run it regardless of the selector's value
+- feat: add `with_selector`, as a decorator to run functions with the result of a selector.
 
 ## Version 0.19.1